backend/requirements.txt
flask==2.3.3
pymongo==4.5.0
flask-pymongo==2.3.0
python-dotenv==1.0.0
flask-jwt-extended==4.5.2
passlib==1.7.4
bcrypt==4.3.0
web3==6.8.0
ipfshttpclient==0.7.0
pytest==7.4.0
flask-cors==4.0.0
gunicorn==21.2.0
mongoengine==0.27.0
python-jose==3.3.0
pydantic==2.3.0
flask-swagger-ui==4.11.1
cryptography==41.0.3
eth-account==0.8.0
py-solc-x==1.1.1
hexbytes==0.3.1
eth-utils==2.2.0
